dash_app/requirements.txt
pip
setuptools
numpy
pandas
dash
plotly
mysql-connector-python